requirements.txt
streamlit==1.41.1
python-dotenv==1.0.1
langchain-community==0.3.15
langchain-text-splitters==0.3.4
langchain-huggingface==0.1.2
langchain-google-genai==3.1.0
pypdf==5.1.0
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
